Use a zero length file as a pseudonym digest value.
The original pseudonym implementation used an all null SHA256
digest as the checksum of a file that was being modeled as a
pseudonym.  This update uses the digest value that is generated
by a zero length file as the digest value for a pseudonym.

This allows a pseudonym map to generate the same security state
value as what is generated by zero'ing those files before running
for example a canister.