Convert model aggregate value to a Buffer object.
The system boot aggregate for a security model has been
previously stored in a 32 byte character array in the state data
of the TSEM object.  This commit converts from using a character
array for this value to a Buffer object.

This provides an additional measure of overflow safety and
prepares for a subsequent update that will return an error only
if an injected aggregate value is different from an aggregate
value that has been pre-programmed into the system.