Rationalize aggregate modeling.
The previous implementation of the TSEM object was not able
address partial security model maps, ie. maps that left the
security model in an unsealed state.  Partial security maps are
useful for refining a security model, for example maps that
contain pseudonym definitions.

This update modifies the TSEM->set_aggregate method to allow the
injection of partial maps with a subsequent change in behavior of
how aggregates are handled.

If a model is sealed any calls to the TSEM->set_aggregate method
simply returns with no action.

If the model is unsealed multiple injections of an aggregate
value will succeed but will cause pertubations in the security
state of the model rather then the generation of an error
condition.