Implement generation of model signing keys.
This commit adds the ability for the sign-model utility to
generate a 2048 bit RSA public/private key pair.  It is designed
to be a convenience function to avoid having to generate a key
using the openssl binary or an equivalent.

The default behavior of the sign-model utility is to sign a
security model map.  Generation of a key is selected with the -G
command-line switch.