Update kernel based security map loading.
This commit updates the quixote utility that manages kernel based
modeling to process security maps that are consistent with the
Sancho modeled Quixote implementations.

The strategy used was to migrate a modified version of the
TSEM->load method into the quixote driver utility along with the
associated private support functions.

Signed models are currently supported, the setup of the security
domain will fail if the model signature is invalid.  The argument
can be made that the kernel should be processing the signature
but that would require a significant update to the Quixote kernel
driver.

Pseudonyms are also not supported given that the current Quixote
kernel driver does not have support for evaluating whether or not
a path description is a pseudonym.  Security event namespaces
have support for an individualed notion of pseudonym definitions
and the lack of pseudonym model support will probably drive an
update to kernel pseudonym processing.