Properly handle cleanup for signed security models.
The quixote implementation was not properly cleaning up state
after a failed security model signature validation.  This commit
addresses the issue.

There is no need to maintain the Buffer objects that hold the
signing key or the signature information data after the signature
verification is complete.  The return value of the request to
verify the signature is retained in a variable that is evaluated
after the objects are released.  Once the release is complete the
return code from the verification function is acted upon.

If the verification function completed without an error the
actual result of the signature verification is processed.

This results in a methodology where the objects containing the
data for signature verification are released as a component of
processing the signature validation request.