Return pseudo-files from a sealed kernel security domain.
Secondary to the architecture used for kernel based Quixote
modeling it is not currently feasible to return security domain
parameters to the quixote-console utility.  Rather then generate
a generic domain failure this update modifies the quixote utility
to return either null measurement values, empty trajectory lists
or a map with a null aggregate value and no security state
points if the domain is sealed.