Initial Nordic SanchoMCU implementation.
This commit implements a framework for building a new SanchoMCU
implementation based on Nordic 52480 micro-controller.

The 52480 is to some degree similar to the STM32l562
micro-controller for which a SanchoMCU implementation has been
done but not yet committed to the Quixote source tree.  One of
the notable characteristics of these parts are support for
hardware based cryptographics primitives such as SHA256 hashing
and asymmetric encryption.

The 52480 part exists in the following two form factors for
development:

NRF52840-DK

NRF52840-DONGLE

The -DK part is the development kit very similar to the
development kits supplied by STmicro and micro-controller
manufacturers.

The -DONGLE part is a small board in a USB form factor of a size
similar to a Yubikey.  This form factor will make it possible to
create Sancho security co-processors that are in a production
form factor.

The -DONGLE part is the part used in the GOOGLE OpenSK project
that provides a development environment for experimenting with
hardware based 2-factor authentication tokens similar to the
Yubikey.

The -DONGLE part is also being used by the FOBNAIL project which
is working to build a co-processor based attestation server for
Secure Launch projects such as Trenchboot.  The FOBNAIL project
envisions that project growing to something that would resemble
what the SanchoMCU implementation already brings to the table,
the ability to provide security guarantees for a Linux
implementation.

The -DK form factor is programmed with a classic USB based
programming architecture while the -DONGLE part is programmed
through the use of a Dynamic Firmware Upgrade (DFU) facility.
Due to anticipated higher flash tolerance and ease of use the
initial SanchoMCU port will be done with the -DK port and then
ported to a build environment for the -DONGLE factor.

Similar to the STmicro port the strategy for the Nordic
implementation will be to have a classic Makefile implement the
build environment.  The Makefile will define include directories
to search in the Nordic Software Development Kit.  VPATH
directives will be used to implementation compilation of the C
and assembly files that are specified for a build.

The initial implementation provides the minimal amount of
infrastructure that is required in order to setup the
micro-controller links and create a linkable image of a program
that does nothing but drop into a continuous loop.