Relocate directory for Quixote control sockets.
Previously the UNIX domain sockets created by Quixote Trust
Orchestration Utitities were rooted in the following location:

/var/run/Quixote

Unfortunately it appears as if modern Linux distributions are
symlinking the /var/run directory to the /run directory.  The
latter of which is being implemented by a temporary RAM based
filesystem.

As a result the control socket directory structure is lost on a
system reboot.

This update moves the root for the directories that hold the
control socket to the following directory:

/var/lib/Quixote/mgmt

Which allows the directory heirarchy created by an installation
to persist through reboot cycles.