Support process id's larger then 16 bits.
This update converts the routines that create the path to the
Quixote management sockets to use dynamic allocation of the PID
identifier with the String->add_sprintf method.

The routines previously used a fixed size buffer for the PID
value which caused them to fail on platforms that are PID's
larger then 16 bits.