Initial USB implementation of Nordic SanchoMCU.
This commit implements the first working USB communications for
the nascent Nordic SanchoMCU implementation.  The implementation
also includes a functional 'console' implementation that will be
useful for future debugging.

One of the major understandings that comes with this commit is
how to properly use the Nordic flash programmer.  After much work
it was learned that this flash programm takes only IHEX as its
input for flashing the micro-controller.  This is in contrast to
the STM programmer that will take an ELF executable as
programming input.

Two USB ACM communications channels are now implemented.  One
that provides the ability for the application to log messages and
one for the actual USB communications over which the Sancho
command protocols will be implemented.

The console is implemented via the external UART while the Sancho
command communications is implemented through the standard USB
controller.  This model will be used moving forward since it will
provide a seamless path to support of the NRF52840-DONGLE which
only has the USB controller.