Introduce Nordic implementation of HurdLib.
The HurdLib implementation is the basic object library that the
Quixote architecture is built on.  This commit implements this
library specific to the Nordic semiconductor implementation of
SanchoMCU.

Consistent with the rest of the Quixote architecture the port of
HurdLib is based on the notion of compiling source from the
reference implementation with platform specific modifications in
order to adapt to the target platform.

This commit introduces changes to the Sancho implementaiton in
order to demonstrate the functionality of the Buffer object which
is the most common predicate object in the library.

Most notably this involves an implementation of the _write()
function in order to support the use of the stdio functions in
HurdLib objects.

This also required a change in the linkage of the SanchoMCU
firmware.  The Nordic development kit had indicated that the
'nano.specs' file should be used to direct linkage.  This
resulted in multiple definition errors when stdio functionality
was used.

The 'nosys.specs' file was introducted to address this problem
and is the linkage specification used for the STmcu Sancho
implementation so consistency is maintained between the two
hardware implementations.