Add error handling function.
The HurdLib library uses the ERR macro in order to provide
backtrace information about an error event.  This macro in turn
calls the Error function to output the description of the event
to an appropriate venue.

This commit implements this function which uses the Nordic
command-line interface to output the error log.