Clarify and fix TTYduct character timing.
The code in the TTYduct->send_Buffer had grown a number of #if
conditionals in an attempt to address transmission reliability to
the SanchoMCU implementations.  Unfortunately those 'hacks'
resulted in code that added no timing whatsoever.

This commit removes the ifdef confusion and establishes a 300
micro-second delay between each transmitted character.  This
needs to be re-examined for correctness across the SanchoMCU
implementations.