Implement initial USB TTY communications.
This commit provides an initial implementation of stable
communications between the host and the Nordic SanchoMCU
implementation.  The Nordic SanchoMCU TTYduct implementation is
based on a native USB ACM-CDC implementation, unlike the STmicro
implementation that was implemented via a ACM-CDC communications
channel supplied by the ST-link debugger implementation.

The latter implementation was used on the STmicro SanchoMCU
implementation secondary to what appears to be issues with the
USB implementation in the STmicro development kit.

The initial framework for the NAAAIM support library was
implemented but only for the purpose of defining the TTYduct
objec timplemenation.  The next commit will provide the NAAAIM
support that will be needed to implement the modeling engine.

The USB communications infrastructure was migrated from the
sancho.c file to the TTYduct object since there is no need to
expose this functionality beyond the TTYduct object.

A significant factor in getting a functional implementation was
to understand the Nordic SDK implementation and the fact that the
app_usbd_event_queue_process() function needs to be called in
order to process the USB event queue.  Without this the event
handlers are not invoked and no USB communication events are
processed.

The TTYduct object needs to be migrated to the actual NAAAIM
library directory but this will be done in combination with a
similar code movement for the STmicro implementation.