Initial Nordic SanchoMCU implementation.
This commit implements the security modeling objects and the
Sancho command interpreter that allows a functional
implementation of a Quixote Trusted Modeling agent on the
NRF52840 Nordic Semiconductor micro-controller.

The implementation of the modeling agent was remarkably straight
forward and largely involved setting up the infrastructure to
build the modeling code from the SecurityModel directory.

Since the Nordic implementation is not using RTOS the command
interpreter was modified so that it would be explicitly called
rather then implicitly invoked as an RTOS thread as is the case
with the STmcu implementation.

The Nordic SanchoMCU implementation has to be manually reset
after each initialization.  This is secondary to some type of
issue with a system reset not re-establishing the CDC-ACM device.