Handle multiple USB connections.
The STm SanchoMCU implementation used a strategy of resetting the
micro-controller when a quixote-mcu instance was terminated.
This is a problematic strategy with the USB implementation since
the reset causes the loss of the CDC-ACM device that in turn
requires another reset cycle in order to re-instantiate the
device.

This update implements support for detecting a CDC-ACM port
closure that in turn causes the termination of the current Sancho
command interpreter instance with a return to waiting for the
next connection.

The detection of the port closure was implemented in the
TTYduct->receive_Buffer method that in turn causes an error
return from the method.  The primary command interpreter loop
detects this and returns from the function after destroying the
current model instance.  The primary event handler then waits for
the detection of an additional port open event to start another
model instance interpreter.

This re-work took a substantial amount of time in order to get
the event handler re-worked, particularly the code that issues
the first read request in response to a port open.  It appears
that there is some type of subtle difference between the first
port open event and subsequent events with respect to setting off
the first read.

The code that handles connection synchronization was also updated
so that a read request is issued each time the receive_handler()
function is called if there is no synchronization character
present.