Fix HurdLib error handling function.
A regression was noted in the Nordic SanchMCU implementation when
the handling of a signed security map with an intential change in
the map was tested.  This caused the SanchoMCU implementation to
hang and become inaccessible to further communications.

The console session would output the same non-printable error
continuously indicating that the NRF log system was somehow being
corrupted and placed into an endless loop.

After an extensive and frustrating debug session the cause was
almost trivial to find and fix, as these things commonly are.

The HurdLib library uses the Error() function supplied by the MCU
implementation to output an error backtrace.  The error function
had been implemented by using snprintf function to print the
formatted error message into a buffer that was then in turn
printed.

The trivial fix was to simply use the NRF_LOG_INFO macro itself
to print out the formatted message.  Which in retrospect should
have been done originally but the initial implementation may have
been influenced by the STmcu SanchoMCU implementation.