Update and improve performance of TTYduct communications.
The performance of the Nordic SanchoMCU implementation was poor
with respect to the amount of time that it takes for a software
cartridge to start.  Work on the NRF52840-DONGLE SanchoMCU
implementation also indicated that the SanchoMCU implementation
was 'hanging' which is typically a manifest of failed
communications between the host and the Sancho implementation.

The host based TTYduct implementation was using a 300
micro-second delay between each character being sent through the
use of the usleep() function.  Decreasing this delay increased
the responsiveness of the SanchoMCU implementation but resulted
in an increasing tendency for the communications to hang.

In order to correct these issues the micro-controller TTYduct
implementation was heavily modified to take advantage of
double-buffering in order to increase transmission throughput.

The following function from the Nordic USB CDC-ACM implementation
is responsible for requesting a read from the host:

app_usbd_cdc_acm_read()

This function schedules the infrastructure to conduct an
asynchronous read over the CDC-ACM port.  The contents of the
buffer supplied through this call are filled in by the interrupt
handler while the primary micro-controller application runs.

The TTYduct implementation was modified to use two separate
buffers to improve performance.  The Receive_Buffer array is
submitted to the app_usbd_acm_read() function to be filled in by
the data stream from the host.

When a receive interrupt (RX OK) is generated the
receive_handler() function copies the contents of the
Receive_Buffer into the Input_Buffer array and then computes the
size of the next block of data to be read.  The next buffer
request is generated from this information and the processing of
the the data in the Input_Buffer is conducted.

This new infrastructure allowed the usleep() calls to be removed
from the transmission of a TTYduct packet.  Tests conducted with
the 'test-sancho' utility indicates that the processing time for
a security interaction event now averages approximately 53.5
milli-seconds compared to a time of approximately 135
mill-seconds under the previous implementation.