Fix regression in TTYduct packet resize after even block packet.
A regression was noted in the Nordic SanchoMCU implementation
that was root caused to a situation where the receipt of TTYduct
packets would stall after a packet with an even number of blocks
was sent.

In this situation the receive_handler() function was incorrectly
setting the size of the next USB read size that was being queued.
Rather then a value of four a read request of one was being
queued.  This in turn was causing the receive_handler() to
set a zero read size that would cause ingress of the next packet
to fail.