Add Nordic SanchoDongle implementation.
This commit implements a SanchoMCU implementation for the
NRF52840-DONGLE part.  The NRF52840-DONGLE is a USB 'fob' form
factor implementation of the NRF52840-DK development board.

The SanchoDongle implementation inherits much of its
functionality from the NRF52840-DK implementation.  A board
specific implementation of the interpreter code is provided in
order to take advantage of the LED's on the board that are used
to provide an indication of the status of the host<->co-processor
implementation.