Fix NRF52840-DONGLE hangs.
The SanchoMCU implementation based on the Nordic Dongle form
factor was reliably hanging when fed a series of security
interaction events.  The hang could also be demonstrated with
interactive exchanges to output the security measurement and
state but less reliably.

This was somewhat puzzling given that the TTYduct implementation
was identical between the NR52840-DK and NRF52840-DONGLE
implementations.

The root cause of the problem was traced to the __WFE() function
that was being called in the data ingress loops in the
TTYduct->receive method.  The __WFE() function issues an ARM
instruction that hints the micro-controller to go into a low
power state until certain system events occur, presumably
interrupts.

Based on this it appears that the __WFE() function has some type
of subtle behavior differences between the two hardware
implementations.

Remove the __WFE() calls in the ingress loops fixes the hanges on
the NRF52840-DONGLE part and does not appear to have any impact
on the NRF52840-DK board.

A __WFE() function was placed at the beginning of the
command-processing loop of the NRF52840-DONGLE implementation,
immediately before the call to the TTYduct->receive call.  This
does not affect the DONGLE implementation and would have the
highest probability of putting the controller into a low power
state when it has the highest probability of waiting for an event
to occur.