Add visual indication of device initialization.
In order to facilitate the visual presentation of the
SanchoDongle implementation the blue LED activity light is
strobed three times after the communications channel was
initialized and before a connection is accepted.

The blue activity LED is extinguished and the light green
connection LED is then illuminated indicating that the device is
ready to accept a connection.

When a connection is accepted the dark green LED indication is
lit.  This LED is extinguished and the BLUE LED illuminated while
a command is being processed.

The change in LED illumination during command processing is
almost imperceptible due to the rapidity of processing most
commands, except for the processing of a security interaction
event.

This processing takes on the order of approximately 50-55
milliseconds in the current implementation.  This causes the
perception of the activity light flickering when a stream of
interaction events is processed.