Implement custom configuration file.
The Nordic development kit requires a configuration file to
specify the large number of parameters that modify the
compilation of code in the development kit.  Each 'project' in
the development kit has its own configuration file.

The NRF52840-DONGLE implementation uses the configuration file
for the usbd_cdc_acm demonstration project as it most closely
aligns with the code needed for the Sancho implementation.  This
configuration file is used to provide definitions for the
NRF52840-DK implementation so customizations are required.

The standard Nordic configuration files have the ability to
override the configurations through an app_config.h file that is
designed to be included in a project implementation.  Defining
the following CPP define:

USE_APP_CONFIG

Causes the standard configuration files to include this config
file at the very start of each file.

This update adds the CPP define to the list of defines that are
passed to all compiled source files in order to enable the use of
a local configuration file.

An app_config.h file is also introduced by this commit to hold
configurations specified to the SanchoDongle implementation.

The initial configuration option included through this file is a
definition which sets the product name for the USB CDC_ACM device
that is implemented to the following:

Enjellic SanchoMCU