Implement reset handling for Nordic SanchoMCU implementations.
The STmicro SanchoMCU implementation uses the 'reset' command at
the end of a connection session to reset the Sancho
implementation.  This is an artifact that needs to be corrected
but the reset command will have utility when the modeling manager
is implemented that will support multiple cartridge sessions in
the MCU.

This update implements a 'connected' variable that maintains the
command processing loop until the 'reset' command is processed.
The reset command sets the connected value to false that
terminates the current command processing session.