Add event hit to NRF52840-DK SanchoMCU implementation.
Work on the NRF52840-DONGLE implementation pointed out a
regression that arose from having the __WFE() hinting function
embedded in the data receive state loops in the
TTYduct->receive_Buffer method.

In the DONGLE implementation a single __WFE instruction was added
to the start of the command processing loop which is the location
in the best position to hint the MCU to enter a low power mode.

This commit adds the instruction to the DK implementation in
order to maintain consistency between the two implementations.