Implement single compilation of regular expressions.
In order to optimize performance on the SanchoMCU implementations
the routines that parse the characteristics of the security
interaction event in the following objects were modified:

COE

Cell

SecurityEvent

The original implementation was a functional but naive
implementation with respect to performance as the regular
expressions that are used to parse the characteristics were
compiled every time the ->parse methods in these objects were
called.

This update introduces a new architecture common across all three
files that compiles all of the needed regular expressions on the
first call to the ->parse method.  Subsequent calls to the method
use the compiled versions of the regular expressions.

On a development machine with a 3.2 GHZ i5-6500 process the first
call to parse a security interaction event takes approximately
0.9 milli-seconds with subsequent calls taking approximately
0.1 to 0.2 milli-seconds.

The primary target for this implementation are the SanchoMCU
implementations where the regular expression parsing cost is a
significant component of the time required to process a security
interaction event.

On a NRF52840-DK development board the first security interaction
event takes approximately 52 milli-seconds to complete.
Subsequent interaction events take on the order of 33.5
milli-seconds to complete.  This reflects approximately a 36%
improvement in performance.