Elimininate hard reset after session close for ST SanchoMCU.
The STmicro SanchoMCU implementation was using a physical reset
to return it to a condition where it could receive an additional
request for a modeling session.  This update implements a session
processing loop that eliminates the need for the hard reset.

This establishes a behavior consistent with the Nordic SanchoMCU
implementation that is based on a USB CDC-ACM communications
model rather then using the CDC-ACM console as the STmicro
implementation does.

This new implementation request that the 'reset' command be
issued from the Quixote implementation in order to maintain
proper connection state.

This change lays the groundworkd for the upcoming Quixote
connection manager that will allow multiple modeling instances to
be supported on a single micro-controller implementation.