Add file mode to test-event utility.
In anticipation of performance tuning on the security interaction
event parsing the ability to time the parsing of an event was
added.

This is selected by the -F command-line option which causes the
utility to read interaction events from standard input.  Each
event is parsed and the time required for the parsing is display.

When run without the -F option the utility runs the standard
regression test on processing of an event.