Modify command-line arguments.
This update modifies the meaning of the command-line arguments to
make them more consistent with the quixote-mcu utility.

The command line switch used to specify the device node used to
access the SanchoMCU implementation was changed from -c to -t.

The -t command-line argument which selected the option to time
the processing of commands was changed to -T.