Introduce STM32L562 SanchoMCU implementation.
This commit provides a SanchoMCU implementation based on the
STmicro STM32L562 micro-controller.  This MCU is focused on the
secure IOT device market and includes TrustZone technology and
hardware support for hashing and asymmetric encryption.

Like the STM32L462 implementation this SanchoMCU implementation
uses the USB console for communications between the host and the
Sancho implementation.  Previous experience with the native USB
controller implementation indicated that there were operational
issues with the USB CDC-ACM implementation.  This has been borne
out with discussions on the ST development list.

Since the STM32L462 and STM32L562 implementations are based on
different MCU implementations they require different platform
switches in order to generate the appropriate binary
instructions.  In order to support the HurdLib and NAAAIM
libraries the implementations for the STM32L562 platform were
included in the same directory heirarchy as the implementation
itself.

The STM32L462 implementation will be modified to follow the same
architecture in subsequent updates.

The current implementation appears to have some type of
regression in that it fails with a HardFault indication just
prior to completing modeling the execution of the Log4j2
application.  The execution of a simple shell cartridge works
properly.

A potential source for the regression may be the limited, 256K,
SRAM size of the STM32L562 platform.  This issue needs further
investigation.  The objective of this commit was to get the base
implementation of the STM32L562 into the source tree.