Initial build framework for SanchoXen implementation.
The next development objective is an implementation of the Sancho
Trusted Modeling Agent inside of a para-virtualized Xen
stubdomain.  The commit initiates the implementation of the build
infrastructure that will be needed to support a standalone Xen
stubdomain.

The current SanchoXen implementation is only a 'hello world' type
program used to prove out the build infrastructure.

This commit implements the following software compoments needed
to implement a stubdomin:

Mini-os.

Mini-os implementation of Newlib C library.

Mini-os implementation of zlib compression library.

Mini-os implemenation of pciutils library.

The build implementation was heavily influenced by the Xen stub
domain implementation.  This build will automatically pull down
implementations of the libraries that are used in the 4.15.2 Xen
implementation.