Implement basic grant/event communications.
This update introduces a prototype for what will be the XenDuct
object that will be used as the object to communicate between a
quixote instance and a SanchoXen stubdomain.  In Xen virtual
devices are implemented with a combination of shared pages of
memory and an 'event-channel' which is a Xen inter-domain
signaling mechanism similar to interrupts.

The design is very simple, assisted by the fact that the 'Duct'
class of objects operate in a write/read framework where there is
no need to support asynchronous writes from the SanchoXen
implementation.  Responses only happen as a result of a command
being send to the SanchoXen instances.

The quixote instance requests that one page of memory be made
available for sharing between itself and the SanchoXen instance.
The quixote instance also creates a bi-directional event channel
that will be used to exchange signals between the two execution
domains.

The quixote instance initiates a connection by writing the
reference to the shared page to the following xenstore path:

/local/domain/MM/backend/SanchoXen/NN/grant-ref

Where MM is the domain number of the Sancho modeling instance and
NN is the the domain number of the domain that is running the
quixote instance.

This is following by writing the number of the event channel to
the following xenstore path:

/local/domain/MM/backend/SanchoXen/NN/event-channel

The first four bytes of the shared memory page are used to
specify that size of the message that is either being sent or
returned.  The remainder of the page is used for the message.

This initial implementation is very primitive and designed as a
proof of concept for the infrastructure.  Tentative plans are
to move this infrastructure into a SanchoXen object.