Fixes to support static quixote-xen implementation in domU.
This update addresses issues that were detected when testing of
the quixote-xen implementation was moved into a domU Linux
instance.

The pthread library was required to be explicitly specified in
order to get the static link to succeed.

The previous Makefile was explicitly using gcc for the compiler
rather then the compiler specified through the CC variable.  This
behavior was reverted since the entire implementation, including
the subordinate libraries, need to be built with a common
compiler in order to avoid linking issues.

If a MUSL based quixote-xen implementation is desired the builder
will need to supply a Xen library package that has been built
with the MUSL library.

The mapping of the shared page was failing since quixote-xen was
specifying the grant reference number in hexadecimal.  Since the
0x prefix was not being applied to the number the conversion of
the string value to a numeric value in SanchoXen was failing.
This was corrected by using a decimal specification of the
grant reference number.

With these changes in place a dom0 based quixote-xen
implementation works properly with a SanchoXen modeling agent in
a stubdomin.  Currently this requires that the xenstore utilities
be used to first write the backend path into the SanchoXen
xenstore entry.  This is followed by setting the permissions on
path to allow the domU instance to write to this path and its
descendants.