Initial implement of mini-os based HurdLib.
This commit adds the HurdLib directory to build a version of the
HurdLib library against the mini-os and newlib implementation.

Following the lead of other architectures the HurdLib build
consists of a Makefile that represents the standard userspace
implementation of library.  The Makefile implements the necessary
defines and directives to build the library against the embedded
implementation.

This commit also introduces the Config.mk include file that will
contain directives specific to the Xen stubdom target
implementation.  Use of this to avoid multiple definitions of
common build infrastructure will be expanded in subsequent
updates.