Implement object based inter-domain communications.
Using the initial grant page/event channel prototype that was
previously developed a XENduct object was implemented for both
the embedded stubdomain environment and standard userspace that
encapsulates the implementation of the communications.

The previous quixote-xen and SanchoXen test harnesses were
updated to use these objects to implement inter-domain
communications.  This was testing with the quixote-xen
implementation running in both dom0 and domU.

The implementation was updated to allow for multiple session
connections to be made for a single object initialization.  This
required the implementation of the following xenstore node:

/local/domain/N/backend/SanchoXen/connect

Which is used by the quixote implementation to signal the
SanchoXen implementation that a connection is being requested.

The XENduct->eof and XENduct->reset methods were implemented in
the stubdomain object in order to facilitate the detection of a
close event by the quixote side of a connection and to reset the
state of the stubdomain object so that it can accept another
connection.