Initial full implementation of SanchoXen.
This update enhances and builds on the XENduct infrastructure to
provide a functional implementation of a SanchoXen
implementation.

This required a number of modifications to the XENduct
infrastructure.  Most notably signalling of when a read was
completed on the quixote side of the XENduct implementation in
order to avoid the sancho implementation from overruning the
information from a previous write from the sancho to the quixote
side.

This issue was noted when the 'show trajectory' command was
tested since the sancho side of the duct would send the
trajectory events in rapid succession.  Without gating by the
receive side the receive shared page would be overrun before the
quixote side complete complete processing of the previous read.

A new utility name test-sancho-xen was implemented in order to
exercise the SanchoXen implementation.  This path was elected
since, in comparison to the MCU implementations, a different
communications object is needed.

The sleep interval in the sancho implementation of the
XENduct->receive_Buffer method was decreased to five
milliseconds.  This was to avoid unnecessarily slowing the
processing of commands that was being caused by the 500
millisecond sleep time.

It should be noted that a proper wait implementation is needed in
order to avoid having to impose these arbitrary wait times.