Full Quixote driver for SanchoXen.
This commit updates the original quixote-xen test harness to a
full Quixote implementation that uses the SanchoXen stub domain.

Since the -d command-line switch in the Quixote series is used to
enable debugging an alternate arguement was needed to specify the
domain id of the SanchoXen implementation.  The -t command-line
argument was re-purposed to specify the numeric domain
identifier.

This is somewhat consistent with the notion that the -t argument
in the SanchoMCU implementation is used to specify the device
node to be used for communicating with the Sancho implementation.
The -t argument can be considered the 'target' trust environment
that the Quixote implementation is to use.