Implement Xen domU setup utility.
Allowing a domU operating system instance to access a SanchoXen
Trusted Modeling Agent (TMA) requires support from the dom0
hypervisor, specifically to setup permissions to allow the dom0
instance to access the xenstore heirarchy for the TMA.

This commit introduces a simple shell based utility that executes
from the dom0 level and setups up the correct 'plumbing'.

The syntax of the command is as follows:

xen-setup Sancho_name dom0_name

Where Sancho_name is replaced with the user-friendly name of the
domain that is implementing the TMA.

dom0_name is replaced with the user-friendly name of the domain
that is to be allowed to connect to the TMA.

The utility uses the userspace xenstore utilities to create the
path in the xenstore key-value heirarchy for the TMA that the
domU instance will be used to signal its connection intent.  The
utilites are also used to set the permissions on the path in
order to allow the dom0 instance to read and write to the
heirarchy.